requirements.txt
numpy
tensorflow
opencv-python
requests
pillow